ast, parser: print statement fully parsed - the power of Pratt parsing
Now a print statement is fully parsed - well a print statement with a
integer literal, other types are not implemented yet.

The Pratt parsing shows itself in the parser with the registering of
parsing functions to token types. The precedence is very basic - there
is only the lowest precedence - but this will change when e.g.
arithmetic operations is implemented.

The node interface in ast got a new String method to pretty print nodes.
This will be helpful when parsing arithemetic operations to see the
grouping by precedence in the future.