ast,compiler: compiling print statement with an integer argument.
The compiler holds a register table for registers that might be used for
intermediate values. This table might be changed since a0-a7 are used
for function arguements in RISC-V. They should probably be changed to
the "t" registers which are temporaries. But for now this is good
enough.
With this, the compiler can allocate and deallocate registers as needed.