compiler: compile print statement with float as argument
This put the float into a label and is loaded from this. There is a way
to compute the float at runtime by generating some assembly code that
plays around with immediate values that shift bits around so they fit
the IEEE 754 format. Then output the assembly code to interpret that
immediate value as float. Though this is a bit hard to do, so maybe
future me can figure that out if necessary :P.