ast: add Expression Statements
With expression statments it will be valid to have expressions in
statement context such as "2\n", "2 + 2\n", or "fn()\n". One could argue
that expression statements such as "2\n" and "2 + 2\n" should not be
allowed as they later in the code gen phase generate uncessary code as
they do not alter any state for the program. This can be checked by the
semantic checker.