ast: fix a bug and fix strings.
There was a bug in the VarStatement.String() as VarStatement does not
always have a value, e.g. "var x int", so a nil pointer dereference
error happens as always tried to call the string method of that value.

The Identifier.String() was missing to output the type of the
Identifier.