ast: add a symbol table to Program node.
With this the Program have the top most symbol table where global
variables and functions are defined. Later when function nodes are
introduced they will themselves have a symbol table which their local
variables are defined within, and have a pointer to the outer symbol
table, e.g. for a normal function the Program symbol table.