resolver: resolve for assignstatement.
Also fix resolving for var statement so is also resolve the rhs, so the
case of "var x int = x" it will check whether x was defined in the
current symbol table. If it succeeds then we check if its a duplicate of
the current symbol table. If that succeeds aswell then the rhs "x" must
come from an outer symbol table aka outer scope and that is valid.