symbol: scope and keep track of num of definitions.
The symbol table holds an outer table, which creates the effect of
scoping. If the outer table is nil, then it is the global scope.

The number of definitions keeps track how many identifiers was declared
in a given scope. This is used to know how much stack space needs to be
reserved for a given block.

On the symbols the Which attribute is used as part of the calculation of
the offset on the stack.