ast,compiler: compile block statements and change register handling.
The register allocation got out of hand. It was kind of a mess to
shuffle around two register tables and keep track of which table to
use. So now it is packed into one register table, and two methods to
allocate for general purpose registers and floating point registers.
Furthermore, getting the name of the registers was annoying, so storing
the register name in the nodes instead of a index into the register
table is way better.