symbol: check recusively outer symbol tables for the identifier.
Before, it only checked its own outer, which means this example would
not yield and error:

{                       // BLOCK 1
    var x int
    {                   // BLOCK 2
        var y int
        {               // BLOCK 3
            var x int
        }
    }
}

as Block 3's outer is Block 2 which did not define an "x". Therefore, we
now recusively walk the outers to check the identifier.

If we have many scopes, defining variables may be slow as we recurse
each time. Although, we could limit of how many nested scopes is allowed
to circumvent this.