types: change types.Type from struct to interface
The struct version of types.Type was limiting the ability to represent
more complex types such as Struct and Func. Therefore, it is changed to
an interface, and new concrete types are created that inherit that
interface.

The Basic struct is for holding basic (primitiv) types such as
int, float, string, and bool.

The Signature struct will be used to represent a func signature.

The Struct struct will be used to represent a struct.