all: new types handling, struct and function introduction.
With the introduction of struct and functions the old type system was
not good enough to describe a more complex type such as a struct and
function.
Furthermore, the parser should not know anything about types, it should
only care about syntax, therefore the types are created in the type
checker (semantic analysis) phase. By doing this, the AST nodes need
more fields that store the syntax information about type assocaiated to
a indentifier. Then in the type checker, these placeholders will be used
for construction a proper type, that is then stored into the nodes in
the tree.
For struct (and certainly also for functions) a new AST node is
introduced that hold the information needed for the type checker to
create the proper type.