resolver,symbol: resolve Selector and Call expression
Furthermore, split up so func statements are defined with DefineFunc and
type statements are defined with DefineType.

The resolution of FuncStatements is to allow for redefining a function
if the previous definition was a function prototype (a function without
a body).

The Type and Function scope are to differentiate them from global and
local scope.