compiler,symbol,types: lot of changes!
TLDR: struct on heap, functions have an epilogue, and global variables
have their own registers allocated.

Structs are allocated on the heap! This is way easier to handle in most
scenarios such passing struct as an argument for function, and return a
struct from a function.
This will also mean that a global struct is just a label holding a
memory address into the heap.

The Func Statement now generates an epilogue that the return statement
will jump to. Before the compiling of Return Statement just emitted "mv
a0, <value register>". This was fine if the Return Statement was the
last instruction in the function although did not work for "early
returns". Therefore, unconditionally jump to the function epilogue fixes
this. With the early return there is a problem with cleaning up the
stack, as the unconditionally jump would jump over all the stack
cleaning instructions. Hence, when we compiling blocks inside a function
we let the functions epilogue clean up any stack space that was
allocated by blocks.

The register table had an extra set of registers "general saved" which
are the registers s{1..11}. The reason for this change was that when
compiling this (omitting the function implementation):

type human struct { name age }
func test() human
var x human = test()

would before become:

.data
x: .dword 0     # label for x struct.
.text
la t0, x        # allocated a temp register for the x label
call test       # call the test function
sd a0, 0(t0)    # store the return value of test into x offset by zero

which looks fine, but the call to test, will not know that the t0 is
still in use, and assume that it can use that value in the function. So
the t0 might be changed in the function and when the "sd a0, 0(t0)" is
run the t0 have the wrong value. Therefore, separating registers
allocated for global variables and local variables, fixes this issue.

New instructions would be:

.data
x: .dword 0     # label for x struct.
.text
la s1, x        # allocated a temp register for the x label
call test       # call the test function
sd a0, 0(s1)    # store the return value of test into x offset by zero