ast,parser,resolver,checker: lot of changes!
ast: with variables being able to have a function type, the type token
Ttoken was not enough since the type checker needed more information
than a simple token. Therefore, Identifier stores TypeNode instead which
a FuncType can be stored into. Furthermore, a BasicType is introduced
which is for any basic type, that just hold the token.

parser: enable the parser to parse variables being of a function type
and also uses the new TypeNodes instead of Ttokens. Furthermore, fixing
a bug in parsing CallExpressions where "somefunc(1)(2)" would parse the
"(2)" as part of the argument for "somefunc". (Although call chaining is
not supported yet in the code gen phase).

resolver: use TypeNode instead of Ttoken.

checker: variables with function as a type are now proper typed with the
information from the TypeNode. Type check that a variable assigned a
function have the same types. The type checker keeps track of the
function that is checked and pass that function to ReturnStatement so
the code gen can use it for the function epilogue.