compiler,symbol: generate code for calling a variable
There is no special code generation to support assigning a variable a
function, as it is just to store the address of the function into the
label. But code for calling avariable is different, as one can not use
the "call" instruction since it takes a label as arguement. To allow
this it needs to use the "jalr" instruction which takes a register.