checker: correct check of identifier with struct type.
Now if an identifier has another identifier as type, then we correctly
check that the identifier is a struct. And we can do that by the
TypeScope since we only give struct types this scope. Of course if we
would allow for more types, then we would have to change this behaviour.